Made Singleton: BiasHashMap which manages the single static sourcebias hashmap. It's populated when u first open app, so now all classes can grab it instead of having to regenerate it every time.
